Log validation requests
This commit introduces logging of validation requests - this enables us
to track usage of the validation tool, and, indirectly, adoption of the
carbon.txt standard.

We do this via a middleware, `LogRequestedDomainMiddleware`, which
intercepts all requests, and adds instrumentation for all requests to
endpoints below the `/api/validate` path.

The middleware will log the endpoint requested, and the `success` status
returned in the response, along with the `url` and `domain` which was
requested, in the case of url validation requests.

We can then create a dasshboard in Grafana to track these particular
events specifically and give an overview of key stats and insights.

Please note that this commit also ups the default log level for the
`carbon_txt` logger to INFO - as we want to see these messages in
Grafana!

I've also refactored the middleware module slightly as we have two very
unrelated middlewares `LogRequestedDomainMiddleware` and
`AddTrailingSlashMiddleware` now live in their own sub-modules of the
`middleware` module, which re-exports both. This way they can live in
their own files, which I think is better for readability and
maintanability.