Log Validation Requests in the database
After some discussion with @mrchrisadams, ([documented on this Trello
ticket](https://trello.com/c/lZfyEHh5/518-carbontxt-api-endpoint-logger)),
we've decided that for now, the simplest way of ensuring we've got
persistent, anonymous logs of validation activity is just to append them
to a database table!

This means we don't have to reconfigure loki / grafana to support a
custom retention period, or think about backing up log data - our
database instance is already persistent and backed up. Plus, it allows
us both to query the logs via sequel, and to display them in grafana,
(as our mariadb instance is already set up as a grafana source).

The API django app already depends on the presence of a db (due to the
admin and session middlewares which are present), but in practice this
is a very soft dependency as django will default to creating a sqlite
database, so it's no great imposition on downstream users of the
validator as an open source package.

The details:

We introduce a ValidationLogEntry model (and migration) to
hold the data, and extend our middleware to create a log entry in the
database on every validation request, alongside logging the information
normally.

I've also refactored all the validation logging functionality into a new
validation_logging django app, and renamed the middleware class for
clarity.