Never follow redirection when validating a URL
When a request for a specific, full carbon.txt file url is requested, we
should not folllow DNS and HTTP delegation logic, and always validate
the specific file requested.

This comes from a discussion with Chris earlier today, in the context of
problems experienced by go.eco (and others) in validating carbon.txt
files hosted behind cloudflare, which adds a Via header which _doesn't_
comform to the standard we expect, and therefore causes an error on
validation.

This causes problems, as their carbon.txt shows as invalid when it
actually is not, for non-obvious reasons which aren't easily addressed
by the user themselves. Instead, much better to show them that their
carbon.txt is in fact valid, while we go about fixing the root cause of
the problem themselves.

It also makes more sense now we have a "validate domain" endpoint which
can handle the logic for delegation and discovery of carbon txts.