Expose delegation method on domain endpoint.
This exposes to the user the mechanism by which a carbon.txt was found
for a domain, in the new `delegation_method` property of the response:
either `null` for a carbon.txt at the domain itself (i.e. not
delegated), `"dns"` for a domain delegated by DNS TXT record, or
l`"http"` for a domain delegated with an HTTP header.

This has involved refactoring the Finder class to return a new
`FinderResult` type which encapsulates both the URL of the the
carbon.txt found and the method by which it was resolved.

One thing we haven't yet resolved is how to express _transitive_ chains
of delegation: when, for instance subdomain.client.com redirects to
client.com, which in turn redirects to provider.com, and where each step
might use a different delegation method. For now, we specificlaly show
the delegation method used in the *first* step, this at least allows
users to debug delegation chains step by step by hand. We should think
some more about what to do in this case, as it's not anything we've ever
seen in the wild, so I don't think it should be a blocker to getting
this out (nor to the LinkedDomain user testing in the GWF provider
portal, which depends on this work.)