Ensure www. subdomain and TLD are normalised in carbon.txt domain resolution
The reasons for this work are outlined in ADR04.

The logic is as follows:

In the event of a DOMAIN lookup for a TLD, or WWW subdomain of a TLD:

- All existing lookup methods are attempted for the requested domain
- If NONE of these methods results in a reachable carbon text, THEN
  the lookup process is repeated for the alternate domain (www in the
  case of a TLD; the TLD in the case of a www subdomain).
- All other lookups (File, URL, requests for other subdomains) are
  unaffected

Implementing this logic requires the `tldextract` library which uses the
public suffix lists to detect whether a domain is a TLD or not. This
downloads and caches the public suffix lists from mozilla, by default
doing this transparently when first used. In the carbon.txt validator
API we elect to explicitly cache the file on boot, so that this process
does not happen during a request/response cycle.